requirements.txt
Django==5.0.4
environs[django]==11.0.0
PyYAML==6.0.1
djangorestframework==3.15.1
PyJWT==2.8.0
factory-boy==3.3.0
pytest-django==4.8.0
psycopg2-binary==2.9.9
requests==2.31.0
redis==5.0.3
celery[redis]==5.3.6
gevent==24.2.1
django-sslserver-v2==1.0
pytest-mock==3.14.0
flake8==7.0.0